requirements.txt
web3>=6.0.0
requests>=2.28.0
python-dotenv>=1.0.0
rich>=13.0.0
eth-abi>=4.0.0
hexbytes>=0.3.0
customtkinter>=5.2.0
